Cleanup dataset upload config component
